Import - basic idea, but missing forms
A basic import page was created, where a file can be selected. However
the node name and description columns and relation columns cannot yet be
provided by the user. Same counts for the mapping of the relations to
the properties, etc.
TO BE WRITTEN indicates where parts have to be completed